[Fix] Fix CLI absolute path rejection, SSAO correctness, shadow cascade degeneracy, and warning cleanup
- Remove absolute path rejection in FileSystem::ValidatePath (needed for CLI test modes)
- Fix SSAO depth comparison to use 3D distance instead of Z-axis only
- Add shadow cascade lookAt up-vector fallback for near-vertical light directions
- Fix RenderScene::Clear() to reset LightData and CameraData
- Fix sign conversion warnings in AssetStore glTF/FBX/OBJ loading code
- Fix unused parameter warnings in GlfwInput and OpenGLContext
- Fix GLsizeiptr conversion warnings in OpenGLBuffer
- Add unit tests for path validation and RenderScene::Clear()

+++ Source/Animation/AnimationClip.h
@@ -163,7 +163,7 @@ public:
             else
                 transform.rotation = glm::normalize(transform.rotation);
 
-            pose.GetLocalTransform(channel.boneIndex) = transform;
+            pose.GetLocalTransform(static_cast<size_t>(channel.boneIndex)) = transform;
         }
 
         pose.ComputeGlobalTransforms(skeleton);
@@ -351,7 +351,7 @@ public:
         if (m_rootMotionEnabled && m_rootMotionBone >= 0 &&
             m_rootMotionBone < static_cast<int32_t>(m_skeleton->GetBoneCount()))
         {
-            auto& boneTransform = m_currentPose.GetGlobalTransform(m_rootMotionBone);
+            auto& boneTransform = m_currentPose.GetGlobalTransform(static_cast<size_t>(m_rootMotionBone));
             if (m_firstRootMotionFrame)
             {
                 m_prevRootPosition = boneTransform.position;
